Introducing labextension
