+++ .gitignore
@@ -6,6 +6,7 @@ node_modules/
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
+jupyterlite_xeus/liteextension
 jupyterlite_xeus/labextension
 # Version file is handled by hatchling
 jupyterlite_xeus/_version.py
